Move extension helpers to `pkg/utils/gardener`
