+++ pkg/operator/controller/extension/extension/runtime/runtime.go
@@ -19,11 +19,11 @@ import (
 
 	v1beta1constants "github.com/gardener/gardener/pkg/apis/core/v1beta1/constants"
 	operatorv1alpha1 "github.com/gardener/gardener/pkg/apis/operator/v1alpha1"
-	"github.com/gardener/gardener/pkg/apis/operator/v1alpha1/helper"
 	resourcesv1alpha1 "github.com/gardener/gardener/pkg/apis/resources/v1alpha1"
 	"github.com/gardener/gardener/pkg/client/kubernetes"
 	"github.com/gardener/gardener/pkg/controllerutils"
 	"github.com/gardener/gardener/pkg/utils"
+	"github.com/gardener/gardener/pkg/utils/gardener"
 	kubernetesutils "github.com/gardener/gardener/pkg/utils/kubernetes"
 	"github.com/gardener/gardener/pkg/utils/managedresources"
 	"github.com/gardener/gardener/pkg/utils/oci"
@@ -91,7 +91,7 @@ func (d *deployer) createOrUpdateResources(ctx context.Context, extension *opera
 		}
 	}
 
-	namespace := helper.ExtensionRuntimeNamespaceName(extension.Name)
+	namespace := gardener.ExtensionRuntimeNamespaceName(extension.Name)
 	if err := d.ensureNamespace(ctx, namespace, extension); err != nil {
 		return fmt.Errorf("failed ensuring namespace %q: %w", namespace, err)
 	}
@@ -101,7 +101,7 @@ func (d *deployer) createOrUpdateResources(ctx context.Context, extension *opera
 		return fmt.Errorf("failed rendering Helm chart %q: %w", extension.Spec.Deployment.ExtensionDeployment.Helm.OCIRepository.GetURL(), err)
 	}
 
-	mrName := helper.ExtensionRuntimeManagedResourceName(extension.Name)
+	mrName := gardener.ExtensionRuntimeManagedResourceName(extension.Name)
 	if err := managedresources.CreateForSeed(ctx, d.runtimeClientSet.Client(), d.gardenNamespace, mrName, false, renderedChart.AsSecretData()); err != nil {
 		return fmt.Errorf("failed creating ManagedResource: %w", err)
 	}
@@ -113,8 +113,8 @@ func (d *deployer) createOrUpdateResources(ctx context.Context, extension *opera
 }
 
 func (d *deployer) deleteResources(ctx context.Context, log logr.Logger, extension *operatorv1alpha1.Extension) error {
-	mrName := helper.ExtensionRuntimeManagedResourceName(extension.Name)
-	namespace := helper.ExtensionRuntimeNamespaceName(extension.Name)
+	mrName := gardener.ExtensionRuntimeManagedResourceName(extension.Name)
+	namespace := gardener.ExtensionRuntimeNamespaceName(extension.Name)
 
 	log.Info("Deleting extension ManagedResource for runtime cluster if present", "managedResource", client.ObjectKey{Name: mrName, Namespace: d.gardenNamespace})
 	if err := client.IgnoreNotFound(managedresources.DeleteForSeed(ctx, d.runtimeClientSet.Client(), d.gardenNamespace, mrName)); err != nil {
