+++ pkg/utils/gardener/extension.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package helper
+package gardener
 
 import (
 	"fmt"
